chore: move configuration steps to configure folder and update the path in files where these are imported
