+++ src/config/api.ts
@@ -1,0 +1,3 @@
+import { API_BASE } from "./env";
+
+export const API_URL = `${API_BASE}/api/v1`;
